github-interaction/requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Deprecated==1.2.13
idna==3.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.10
wrapt==1.14.1
